Refactor pick_n_place_demo
Move PickAndPlaceDomain into src
and just import it in the nodes.
Fixes #17.
